+++ src/components/icons/icons.tsx
@@ -98,7 +98,9 @@ export const IconPlayCircle = createIconBase({
 });
 
 export const IconGridView = createIconBase({
-  paths: fill("M3 3v8h8V3H3zm6 6H5V5h4v4zm-6 4v8h8v-8H3zm6 6H5v-4h4v4zm4-16v8h8V3h-8zm6 6h-4V5h4v4zm-6 4v8h8v-8h-8zm6 6h-4v-4h4v4z"),
+  paths: fill(
+    "M3,3v8h8V3H3z M9,9H5V5h4V9z M3,13v8h8v-8H3z M9,19H5v-4h4V19z M13,3v8h8V3H13z M19,9h-4V5h4V9z M13,13v8h8v-8H13z M19,19h-4v-4h4V19z",
+  ),
 });
 
 export const IconListView = createIconBase({